fix(suite-desktop): unify app restart logic
